requirements.txt
pandas>=2.0,<3.0
pyodbc>=5.0,<6.0
requests>=2.31,<3.0
azure-identity>=1.15,<2.0
azure-storage-file-datalake>=12.14,<13.0
python-dotenv>=1.0,<2.0
pydantic>=2.5,<3.0
pydantic-settings>=2.1,<3.0
tenacity>=8.2,<9.0
pyarrow>=14.0,<18.0
simple-salesforce>=1.12,<2.0

# Dev / Testing
pytest>=7.0,<9.0
pytest-cov>=4.0,<6.0
